Farcaster image support
Farcaster takes a link to the image so I also implemented a function
that uploads your image to an S3 bucket. It requires the user to manage
the aws cli which isn't ideal but is fine for now. Should implement
Imgur. I changed the farcaster python script to only post to farcaster
with the looping handled in index.js so the python file doesn't have to
manage getting the permalinks itself.